fix(rdf-writer): kill 4 dead Fuseki channels found in phase-1 audit
Live-verified as part of the recall/Falkor-cutover Phase 1 audit
(docs/superpowers/specs/2026-07-17-recall-rdf-writer-falkor-cutover-phase2-spec.md):
four rdf-writer subscriptions/dispatch branches have zero real live traffic
or a real durable consumer already exists elsewhere, matching the
2026-07-15 drive-audit and cognition/metacog (PR #1155) kill precedent.

- orion:rdf:worker (cortex.worker.rdf_build / CognitiveStepExecution):
  zero real producers anywhere in the repo despite channels.yaml claiming
  orion-cortex-exec -- the one other hit was self_study.py referencing the
  channel name in a self-knowledge catalog string, not a publish call.
- orion:cortex:telemetry: zero producers anywhere, not even registered in
  channels.yaml.
- orion:memory:identity:snapshot / orion:memory:goals:proposed: live
  Fuseki query found zero graphs matching autonomy/identity/goals ever
  recorded, despite a real producer existing (orion/spark/concept_induction).
  Already gated inert via RDF_SKIP_KINDS since a 2026-07-16 disk incident
  (83M of 85M store triples, no retention) -- this removes the subscription
  structurally instead of relying on the runtime skip flag. Postgres
  (orion-self-state-runtime's SelfStateRuntimeStore) already durably owns
  identity snapshots; orion-substrate-runtime's goal_context_listener.py
  already consumes goal proposals live.
- orion:world_pulse:graph:upsert: WORLD_PULSE_GRAPH_ENABLED was false in
  the live .env (fully inert), graph shape was 3 flat literals per digest
  item with no edges, and world-pulse's richer claim/event/entity emit
  channels already reach Postgres via orion-sql-writer.

Deleted app/autonomy.py entirely (only served the two killed kinds).
Deleted test_autonomy_materialization.py and test_world_pulse_graph_gates.py
(tested only removed code); added test_dead_channel_kill.py with
channel-not-subscribed + quiet-no-op dispatch regression guards for all 4;
moved the still-live SocialRoomTurn test into
test_service_rdf_store_integration.py.

Also cleaned RDF_SKIP_KINDS (identity/goals entries now dead since the
kind can never reach the skip check) and fixed channels.yaml registry
accuracy for all 4 channels. orion-world-pulse's producer side (still
publishes into a now-consumerless channel) flagged as an out-of-scope
follow-up, not touched here.

Env parity: same pre-existing 17-key .env_example/docker-compose.yml gap
before and after (PROJECT/NET/Fuseki JVM tuning, unrelated) -- no new gap.
Local .env synced directly (dangling keys removed, RDF_SKIP_KINDS trimmed).
