+++ services/orion-rdf-writer/.env_example
@@ -26,16 +26,21 @@ CHANNEL_EVENTS_COLLAPSE=orion:collapse:intake
 CHANNEL_EVENTS_TAGGED=orion:tags:enriched
 CHANNEL_EVENTS_TAGGED_CHAT=orion:tags:chat:enriched
 CHANNEL_CORE_EVENTS=orion:core:events
-CHANNEL_WORKER_RDF=orion:rdf:worker
-CORTEX_LOG_CHANNEL=orion:cortex:telemetry
+# CHANNEL_WORKER_RDF removed 2026-07-18: orion:rdf:worker has zero real
+# producers anywhere in the repo despite the channel registry claiming one;
+# rdf-writer no longer subscribes or handles the kind.
+# CORTEX_LOG_CHANNEL removed 2026-07-18: orion:cortex:telemetry has zero
+# producers anywhere, not even registered in channels.yaml.
 CHANNEL_COGNITION_TRACE_PUB=orion:cognition:trace
 CHANNEL_CHAT_HISTORY_TURN=orion:chat:history:turn
 CHANNEL_CHAT_HISTORY_LOG=orion:chat:history:log
-CHANNEL_MEMORY_IDENTITY_SNAPSHOT=orion:memory:identity:snapshot
+# CHANNEL_MEMORY_IDENTITY_SNAPSHOT / CHANNEL_MEMORY_GOALS_PROPOSED removed
+# 2026-07-18: zero live Fuseki triples ever recorded despite real producers
+# existing; Postgres/orion-substrate-runtime already durably own this data.
 # CHANNEL_MEMORY_DRIVES_AUDIT removed 2026-07-15: drive audits are Postgres-only
 # (`drive_audits` via orion-sql-writer); rdf-writer no longer subscribes or handles the kind.
-CHANNEL_MEMORY_GOALS_PROPOSED=orion:memory:goals:proposed
-CHANNEL_WORLD_PULSE_GRAPH=orion:world_pulse:graph:upsert
+# CHANNEL_WORLD_PULSE_GRAPH removed 2026-07-18: WORLD_PULSE_GRAPH_ENABLED was
+# false live (fully inert); rdf-writer no longer subscribes or handles the kind.
 
 # --- Bus Publish Channels ---
 CHANNEL_RDF_CONFIRM=orion:rdf:confirm
@@ -117,9 +122,12 @@ FUSEKI_JVM_ARGS=-Xms${FUSEKI_JVM_XMS} -Xmx${FUSEKI_JVM_XMX} ${FUSEKI_JVM_GC_OPTS
 
 # --- Runtime ---
 LOG_LEVEL=INFO
-# Skip autonomy theater materialization (goals/identity snapshots) while chat stance graph reads are off.
 # memory.drives.audit.v1 is no longer listed: the handler was deleted 2026-07-15 (Postgres-only via orion-sql-writer).
-RDF_SKIP_KINDS=chat.history.message.v1,memory.identity.snapshot.v1,memory.goals.proposed.v1
+# memory.identity.snapshot.v1 / memory.goals.proposed.v1 are no longer listed 2026-07-18:
+# the channel subscriptions and handlers were removed entirely (not subscribed at all now,
+# not just skipped) -- same shape as the drives-audit kill. Listing them here would be dead
+# config since the kind can never reach this skip check anymore.
+RDF_SKIP_KINDS=chat.history.message.v1
 RDF_SKIP_REJECTED=true
 RDF_DURABLE_ONLY=false
 
@@ -146,8 +154,11 @@ RDF_RETENTION_TIMEOUT_SEC=3600
 # slow one-artifact-at-a-time) — see docs/superpowers/pr-reports/2026-07-16-fuseki-disk-cleanup-pr.md.
 RDF_RETENTION_POLICIES=[{"graph":"http://conjourney.net/graph/autonomy/identity","max_artifacts":500,"batch_size":500},{"graph":"http://conjourney.net/graph/autonomy/goals","max_artifacts":2000,"batch_size":500},{"graph":"http://conjourney.net/graph/autonomy/drives","max_artifacts":1,"batch_size":500}]
 
-WORLD_PULSE_GRAPH_ENABLED=false
-WORLD_PULSE_GRAPH_DRY_RUN=true
-WORLD_PULSE_GRAPH_REQUIRE_POLICY_STAMP=true
+# WORLD_PULSE_GRAPH_ENABLED / WORLD_PULSE_GRAPH_DRY_RUN /
+# WORLD_PULSE_GRAPH_REQUIRE_POLICY_STAMP removed 2026-07-18: the
+# world.pulse.graph.upsert.v1 handler and CHANNEL_WORLD_PULSE_GRAPH
+# subscription were removed entirely (was already disabled live, dry_run
+# by default even when enabled). Settings fields deleted -- nothing reads
+# them anymore.
 RETRY_LIMIT=3
 RETRY_INTERVAL=2
